+++ ChangeLog.md
@@ -16,6 +16,7 @@
 12. （BUG）生成 subtree 时没有连接所有节点
 13. （BUG）搜索节点时没有正确处理中间状态（unstable state）
 14. （BUG）（13衍生）获取 key 偏移时错误理解小端字节序
+15. （FIX）解决 border node 的 lower key 问题，但是降低了性能，同时增加了内存使用。A necessary evil。